# routes/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IResponse": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double",
						"description": "This is the code of response you consult the code list for known what this code meaning"
					},
					"data": {
						"description": "This is the data (Optionnal)"
					},
					"message": {
						"type": "string",
						"description": "This is the message of the request (Optionnal)"
					},
					"token": {
						"type": "string",
						"description": "This is the token givin on login (Optionnal)"
					}
				},
				"required": [
					"code",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "token",
				"in": "header",
				"scopes": {
					"root": "can do",
					"admin": "can do",
					"user": "can do"
				}
			}
		}
	},
	"info": {
		"title": "Node_Kernel",
		"version": "1.0.0",
		"description": "Node Js Kernel functionnal with Typescript and Sequelize",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "D Smart"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/index/supercontroller": {
			"post": {
				"operationId": "Sample",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"description": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}